flask/requirements.txt
Flask
requests
googlemaps
os
dotenv
google.generativeai
openai